Move collection to a new file.
