+++ authority/provisioner/provisioner.go
@@ -3,7 +3,6 @@ package provisioner
 import (
 	"encoding/json"
 	"strings"
-	"sync"
 
 	"github.com/pkg/errors"
 )
@@ -44,6 +43,11 @@ func (p *Provisioner) ID() string {
 	return p.base.ID()
 }
 
+// Type return the provisioners type.
+func (p *Provisioner) Type() Type {
+	return p.typ
+}
+
 // Init initializes the base provisioner with the given claims.
 func (p *Provisioner) Init(claims *Claims) error {
 	return p.base.Init(claims)
@@ -70,48 +74,16 @@ func (p *Provisioner) UnmarshalJSON(data []byte) error {
 
 	switch strings.ToLower(typ.Type) {
 	case "jwt":
+		p.typ = TypeJWK
 		p.base = &JWT{}
 	case "oidc":
+		p.typ = TypeOIDC
 		p.base = &OIDC{}
 	default:
 		return errors.New("provisioner type not supported")
 	}
 	if err := json.Unmarshal(data, &p.base); err != nil {
 		return errors.Errorf("error unmarshalling provisioner")
-	}
-	return nil
-}
-
-// Collection is a memory map of provisioners.
-type Collection struct {
-	byID *sync.Map
-}
-
-// NewCollection initializes a collection of provisioners.
-func NewCollection() *Collection {
-	return &Collection{
-		byID: new(sync.Map),
-	}
-}
-
-// Load a provisioner by the ID.
-func (c *Collection) Load(id string) (*Provisioner, bool) {
-	i, ok := c.byID.Load(id)
-	if !ok {
-		return nil, false
-	}
-	p, ok := i.(*Provisioner)
-	if !ok {
-		return nil, false
-	}
-	return p, true
-}
-
-// Store adds a provisioner to the collection, it makes sure two provisioner
-// does not have the same ID.
-func (c *Collection) Store(p *Provisioner) error {
-	if _, loaded := c.byID.LoadOrStore(p.ID(), p); loaded == false {
-		return errors.New("cannot add multiple provisioners with the same id")
 	}
 	return nil
 }
